requirement.txt
# Discord Bot Dependencies
discord.py>=2.3.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.8.0

# AI/LLM Integration
google-genai>=0.3.0

# Data Validation
pydantic>=2.0.0